requirements.txt
absl-py==1.2.0
altair==4.2.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
contourpy==1.0.5
cryptography==38.0.1
cupy-cuda11x
cycler==0.11.0
decorator==5.1.1
dnspython==2.2.1
docopt==0.6.2
easydict==1.10
entrypoints==0.4
executing==1.1.0
fastrlock==0.8
fonttools==4.37.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-crc32c==1.5.0
grpcio>=1.60.1
idna==3.4
imageio==2.22.2
importlib-metadata==4.12.0
importlib-resources==5.9.0
imutils==0.5.4
ipython==7.22.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
kiwisolver==1.4.4
llvmlite>=0.44.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
natsort==8.2.0
netifaces==0.11.0
numba==0.56.2
numbalsoda==0.2.1
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.5.5.62
packaging==21.3
pandas==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0b3
pyee==9.0.4
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
scipy==1.9.1
seaborn==0.12.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
stack-data==0.5.1
streamlit==1.13.0
streamlit-cropper==0.2.0
streamlit-player==0.1.5
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
toml==0.10.2
toolz==0.12.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu113
tornado==6.3.3
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
Werkzeug==2.2.2
yarg==0.1.9
zipp==3.8.1
